Back/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
arelle-release==2.35.9
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
billiard==4.2.1
bleach==6.2.0
blinker==1.6.3
Brotli==1.1.0
celery==5.3.0
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.0
cycler==0.12.1
dart-fss==0.4.10
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==7.0.0
Django==4.2.16
django-allauth==65.2.0
django-cors-headers==4.6.0
django-extensions==3.2.3
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
fake-useragent==1.5.1
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.0.0
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
fonttools==4.55.0
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.5.0
gevent==23.9.1
geventhttpclient==2.0.11
greenlet==3.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.27.2
idna==3.4
importlib-metadata==6.8.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.18.1
ipython-genutils==0.2.0
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.2
jiter==0.7.1
json5==0.9.28
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_contrib_core==0.4.2
jupyter_contrib_nbextensions==0.7.0
jupyter_core==5.7.2
jupyter_nbextensions_configurator==0.6.4
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.7
kombu==5.3.7
locust==2.17.0
lxml==5.3.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.7
mistune==3.0.2
msgpack==1.0.7
multidict==6.1.0
multitasking==0.0.11
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.1
openai==0.28.0
OpenDartReader==0.2.3
openpyxl==3.1.5
overrides==7.7.0
packaging==24.2
pandas==2.1.1
pandocfilters==1.5.1
parso==0.8.4
peewee==3.17.8
pillow==11.0.0
platformdirs==4.3.6
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==5.9.6
pure_eval==0.2.3
pycparser==2.21
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
python-docx==1.1.2
python-json-logger==2.0.7
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==25.1.1
redis==5.0.0
referencing==0.35.1
regex==2024.11.6
requests==2.31.0
requests-file==2.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roundrobin==0.0.4
rpds-py==0.21.0
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.4.4
stack-data==0.6.3
termcolor==2.3.0
terminado==0.18.1
tinycss2==1.4.0
tomli==2.1.0
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.1
xmltodict==0.14.2
yarl==1.18.0
yaspin==3.1.0
yfinance==0.2.50
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
